mcp_server/requirements.txt
# MCP Server dependencies
# SDK oficial de Model Context Protocol
mcp[cli]>=1.2.0

# HTTP client asíncrono (usado en las tools)
httpx>=0.27.0

# Requests síncrono (fallback y utilidades)
requests>=2.31.0

# Hot-reload en desarrollo (equivalente a gunicorn --reload)
watchfiles>=0.21.0

# JWT signing para service tokens (sin usuario/contraseña)
PyJWT>=2.8.0

